fix numpy deprecation

+++ bbox/bbox3d.py
@@ -86,7 +86,7 @@ class BBox3D:
         if not np.isscalar(x):
             raise ValueError("Value should be a scalar")
         else:  # x is a scalar so we check for numeric type
-            if not isinstance(x, (np.float, np.int)):
+            if not isinstance(x, (float, int)):
                 raise TypeError("Value needs to be either a float or an int")
         return x
